requirements.txt
PyQt5
scapy
requests
python-nmap
geopy